renamed validator.ts
